requirements.txt
beautifulsoup4
requests
boto3
discord.py
